// frontend/packages/web/src/enums/DashboardWidgetType.ts
export enum DashboardWidgetType {
  BillPaymentsStatusPercentage = 'BILL_PAYMENTS_STATUS_PERCENTAGE',
  ContractsActiveExpirationsList = 'ACTIVE_CONTRACTS_EXPIRATIONS_LIST',
  ContractsPassiveExpirationsList = 'PASSIVE_CONTRACTS_EXPIRATIONS_LIST',
  ContractsRecap = 'CONTRACTS_RECAP',
  EstateMap = 'ESTATE_MAP',
  EstateOccupationStatusPercentage = 'ESTATE_OCCUPATION_STATUS_PERCENTAGE',
  EstateUnitTypesPercentage = 'ESTATE_UNIT_TYPES_PERCENTAGE',
  IncomeOutcomeComparisonChart = 'INCOME_OUTCOME_COMPARISON_CHART',
  NotificationsList = 'NOTIFICATIONS_LIST',
  RealEstateRecap = 'REAL_ESTATE_RECAP',
  TicketsChecklistExpirationsList = 'CHECKLIST_TICKETS_EXPIRATIONS_LIST',
  TicketsCompletedChart = 'TICKETS_COMPLETED_CHART',
  TicketsCountChart = 'TICKETS_COUNT_CHART',
  TicketsExcludedFromMaintenanceContractChart = 'TICKETS_EXCLUDED_FROM_MAINTENANCE_CONTRACT_CHART',
  TicketsExcludedFromMaintenanceContractExpensesChart = 'TICKETS_EXCLUDED_FROM_MAINTENANCE_CONTRACT_EXPENSES_CHART',
  TicketsExcludedFromMaintenanceContractExpensesPercentage = 'TICKETS_EXCLUDED_FROM_MAINTENANCE_CONTRACT_EXPENSES_PERCENTAGE',
  TicketsExcludedFromMaintenanceContractPercentage = 'TICKETS_EXCLUDED_FROM_MAINTENANCE_CONTRACT_PERCENTAGE',
  TicketsExcludedFromMaintenanceContractTotalExpensesCount = 'TICKETS_EXCLUDED_FROM_MAINTENANCE_CONTRACT_TOTAL_EXPENSES_COUNT',
  TicketsIssueExpirationsList = 'ISSUE_TICKETS_EXPIRATIONS_LIST',
  TicketsMandatoryByLawExpirationsList = 'MANDATORY_BY_LAW_TICKETS_EXPIRATIONS_LIST',
  TicketsMandatoryByLawExpired = 'TICKETS_MANDATORY_BY_LAW_EXPIRED',
  TicketsMandatoryByLawList = 'TICKETS_MANDATORY_BY_LAW_LIST',
  TicketsMandatoryByLawStatusPercentage = 'TICKETS_MANDATORY_BY_LAW_STATUS_PERCENTAGE',
  TicketsMandatoryByLawTypePercentage = 'TICKETS_MANDATORY_BY_LAW_TYPE_PERCENTAGE',
  TicketsPreventativeExpired = 'TICKETS_PREVENTATIVE_EXPIRED',
  TicketsPriorityPercentage = 'TICKETS_PRIORITY_PERCENTAGE',
  TicketsSLARespectingPercentage = 'TICKETS_SLA_RESPECTING_PERCENTAGE',
  TicketsStatusAverageResolutionDuration = 'TICKETS_STATUS_CHART_AVERAGE_RESOLUTION_DURATION',
  TicketsStatusChart = 'TICKETS_STATUS_CHART',
  TicketsStatusCount = 'TICKETS_STATUS_COUNT',
  TicketsStatusPercentage = 'TICKETS_STATUS_PERCENTAGE',
  TicketsTotalExpensesChart = 'TICKETS_TOTAL_EXPENSES_CHART',
  TicketsTypeChart = 'TICKETS_TYPE_CHART',
  TicketsTypeList = 'TICKETS_TYPE_LIST',
  UsageTypesPercentage = 'USAGE_TYPES_PERCENTAGE',
}